jarvis 1/requirements.txt
deepgram-sdk
pygame
edge_tts
selenium
python-dotenv
speechrecognition